erts: Extract out the to_chars used by ryu form STL
We place it in a separate file in order to move out
the license information and to make it easier to pick
updates.
