+++ erts/emulator/ryu/vendor.info
@@ -12,11 +12,35 @@
     "copyrightText": "Copyright 2018 Ulf Adams",
     "downloadLocation": "https://github.com/ulfjack/ryu",
     "homepage": "https://github.com/ulfjack/ryu",
-    "licenseDeclared": "Apache-2.0 OR BSL-1.0",
+    "licenseDeclared": "Apache-2.0",
     "name": "ryu",
     "versionInfo": "844864ac213bdbf1fb57e6f51c653b3d90af0937",
-    "path": "./erts/emulator/ryu",
+    "path": ["./erts/emulator/ryu/common.h",
+             "./erts/emulator/ryu/d2s.c",
+             "./erts/emulator/ryu/d2s_full_table.h",
+             "./erts/emulator/ryu/d2s_intrinsics.h",
+             "./erts/emulator/ryu/digit_table.h",
+             "./erts/emulator/ryu/ryu.h",
+             "./erts/emulator/ryu/ryu.mk",
+             "./erts/emulator/ryu/README.ryu_update.md"],
     "supplier": "Person: Ulf Adams",
-    "purl": "pkg:github/ulfjack/ryu"
+    "purl": "pkg:github/ulfjack/ryu#ryu",
+    "update": "./erts/emulator/ryu/update.sh"
+  },
+  {
+    "ID": "ryu-to_chars",
+    "description": "to_chars from STL ryu",
+    "copyrightText": "Copyright 2018 Ulf Adams\nCopyright (c) Microsoft Corporation.",
+    "downloadLocation": "https://github.com/microsoft/STL",
+    "homepage": "https://github.com/microsoft/STL",
+    "licenseDeclared": "Apache-2.0 WITH LLVM-exception AND BSL-1.0",
+    "name": "ryu_to_chars",
+    "versionInfo": "vs-2019-16.10",
+    "path": [
+      "./erts/emulator/ryu/to_chars.h"
+    ],
+    "supplier": "Organization: Microsoft Corporation",
+    "purl": "pkg:github/microsoft/STL#stl/inc/xcharconv_ryu.h",
+    "update": "./erts/emulator/ryu/update.sh"
   }
 ]
